Request midi in score model
In previous commits, we requested the MIDI file associated with a score
model in the score route. However, we may need to make this request in
contexts outside the score route. This commit moves this request into
the score model so that we can get score.midi in any context where
score.fileUrl is defined.