Build S3 direct upload policy in service:s3
In previous commits, we requested from the server all information
necessary to upload a file directly to s3. However, the most of this
information was already available client-side; the only additional
field we need is a signature for the given policy. This commit builds
the S3 direct upload policy locally and sends it to the server to
acquire a signature. This approach affords more flexibility and
visibility regarding how the direct upload policy works, and also
allows our server to be more RESTful.