Display only five seconds of notes in component:player-piano
In previous commits, the player piano template rendered DOM for every
note in a displayed MIDI file. This excess DOM causes browsers to lag,
resulting in a poor user experience. We need to display only those
notes visible within the current viewport, and so this commit
approximates that by displaying only those notes on during the next
five seconds.